+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2.10.4-wip.4
+- :bulb: Included abilities field for each user to messages#index response
+- :arrow_up: Updated .pryrc
+
 ### 2.10.4-wip.3
 - :arrow_up: Removed unnecessary code
 - :hammer: Fixed rack middleware for api documentation